pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gpx-poi-enricher"
version = "0.1.0"
description = "Enrich GPX tracks with Points of Interest from OpenStreetMap via configurable YAML profiles"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["gpx", "openstreetmap", "overpass", "poi", "waypoints", "travel", "geocoding"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Utilities",
]
dependencies = [
    "requests>=2.28",
    "gpxpy>=1.6",
    "PyYAML>=6.0",
]

[project.scripts]
gpx-poi-enricher = "gpx_poi_enricher.cli:main"
gpx-split-waypoints = "gpx_poi_enricher.split_cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "responses>=0.25",   # HTTP mocking for tests
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/devmarkusb/gpx-poi-enricher"
Repository = "https://github.com/devmarkusb/gpx-poi-enricher"
Issues = "https://github.com/devmarkusb/gpx-poi-enricher/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/gpx_poi_enricher"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.coverage.run]
source = ["src/gpx_poi_enricher"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
